AOC.Challenge.Day17

Description

Documentation

day17a :: Set Point :~> Int

day17b :: Set Point :~> Int

runDay17

Arguments

:: Bool
cache thunk between runs

-> Int
number of steps

-> Int
dimensions

-> Set Point
points

-> [IntMap IntSet]
steps

ixPascal

Arguments

:: Int
dimension

-> Int

-> [Int]

pascalIx :: [Int] -> Int

encRun :: Int -> [Int] -> [Int]

pascalVecRunIx :: Vector Int -> Int

vecRunIxPascal

Arguments

:: Int
dimension

-> Int
maximum

-> Int
number

-> [Int]
run

pascalTable :: Int -> Int -> [[Int]]

oldNeighborWeights

Arguments

:: Int
dimension

-> Int
maximum

-> Vector (IntMap NCount)

vecRunNeighbs

Arguments

:: [[Int]]
pascal table

-> Int

-> [(Int, NCount)]

Streaming/constant space enumerate all neighbor and multiplicities

vecRunNeighbs_ :: Vector Int -> [(Vector Int, NCount)]

Reference implementation for vecRunNeighbs, which takes and returns actual vec run neighbors

validNCount :: NCount -> Maybe Bool

neighborWeights

Arguments

:: Int
dimension

-> Int
maximum

-> Vector (IntMap NCount)

Build up all the weights for quick reference comparison

finalWeight

Arguments

:: (Num a, Ord a)

=> Int
dim

-> [a]

-> Int