AOC.Challenge.Day18

Description

Day 18. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

data Expr

Constructors

Mul Expr Expr

Add Expr Expr

Lit Int

parseLit :: CharParser Expr

parsePrim :: CharParser Expr

parseExpr :: CharParser Expr

tok :: forall f e s a. (MonadParsec e s f, Token s ~ Char) => f a -> f a

eval :: Expr -> Int

day18a :: _ :~> _

parsePrim2 :: CharParser Expr

parsePrim22 :: CharParser Expr

parseExpr2 :: CharParser Expr

day18b :: _ :~> _