AOC.Common

Description

Common functions for solutions

Loops and searches

iterateMaybe :: (a -> Maybe a) -> a -> [a]

Iterate until a Nothing is produced

loopMaybe :: (a -> Maybe a) -> a -> a

Apply function until Nothing is produced, and return last produced value.

loopMaybeM :: Monad m => (a -> m (Maybe a)) -> a -> m a

Apply monadic function until Nothing is produced, and return last produced value.

loopEither :: (a -> Either r a) -> a -> r

Apply function until a Left.

firstJust :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Like find, but instead of taking an a -> Bool, takes an a -> Maybe b and returns the first success.

(!!!) :: [a] -> Int -> a

Strict (!!)

(!?) :: [a] -> Int -> Maybe a

drop' :: Int -> [a] -> [a]

Strict drop

dup :: a -> (a, a)

A tuple of the same item twice

scanlT :: Traversable t => (b -> a -> b) -> b -> t a -> t b

scanl generalized to all Traversable.

scanrT :: Traversable t => (a -> b -> b) -> b -> t a -> t b

scanr generalized to all Traversable.

firstRepeated :: Ord a => [a] -> Maybe a

Lazily find the first repeated item.

firstRepeatedBy :: Ord a => (b -> a) -> [b] -> Maybe b

Lazily find the first repeated projection.

fixedPoint :: Eq a => (a -> a) -> a -> a

Repeat a function until you get the same result twice.

floodFill

Arguments

:: Ord a

=> (a -> Set a)
Expansion (be sure to limit allowed points)

-> Set a
Start points

-> Set a
Flood filled

Flood fill from a starting set

floodFillCount

Arguments

:: Ord a

=> (a -> Set a)
Expansion (be sure to limit allowed points)

-> Set a
Start points

-> (Int, Set a)
Flood filled, with count of number of steps

Flood fill from a starting set, counting the number of steps

countTrue :: Foldable f => (a -> Bool) -> f a -> Int

Count the number of items in a container where the predicate is true.

Lists

freqs :: (Foldable f, Ord a) => f a -> Map a Int

Build a frequency map

lookupFreq :: Ord a => a -> Map a Int -> Int

Look up a count from a frequency map, defaulting to zero if item is not foudn

freqList :: (Foldable f, Ord a) => f a -> [(Int, a)]

Build a list of descending frequencies. Ties are sorted.

revFreq :: (Foldable f, Ord a) => f a -> IntMap (NESet a)

Build a reverse frequency map

perturbations :: Traversable f => (a -> [a]) -> f a -> [f a]

Collect all possible single-item perturbations from a given perturbing function.

perturbations (\i -> [i - 1, i + 1]) [0,10,100]

[ [-1,10,100]

, [ 1,10,100] , [ 0, 9,100] , [ 0,11,100] , [ 0,10, 99] , [ 0,10,101] ]

perturbationsBy :: Conjoined p => Over p (Bazaar p a a) s t a a -> (a -> [a]) -> s -> [t]

Collect all possible single-item perturbations from a given perturbing function.

perturbations (\i -> [i - 1, i + 1]) [0,10,100]

[ [-1,10,100]

, [ 1,10,100] , [ 0, 9,100] , [ 0,11,100] , [ 0,10, 99] , [ 0,10,101] ]

select :: [a] -> [(a, [a])]

each item paired with the list not including that item

slidingWindows :: Int -> [a] -> [Seq a]

sliding windows of a given length

sortedSlidingWindows :: forall k v. Ord k => Int -> [(k, v)] -> [OrdPSQ k Int v]

sorted windows of a given length

sortedSlidingWindowsInt :: forall v. Int -> [(Int, v)] -> [IntPSQ Int v]

sorted windows of a given length

clearOut :: (Char -> Bool) -> String -> String

Clear out characters not matching a predicate

foldMapPar :: Monoid b => (a -> b) -> [a] -> b

foldMap, but in parallel.

foldMapPar1 :: Semigroup b => (a -> b) -> NonEmpty a -> b

foldMap1, but in parallel.

foldMapParChunk

Arguments

:: (NFData m, Monoid m)

=> Int
chunk size

-> (a -> m)

-> [a]

-> m

meanVar :: Fractional a => Fold a (a, a)

Fold for computing mean and variance

maximumVal :: Ord b => Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the maximum value in the map

maximumValBy :: (b -> b -> Ordering) -> Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the maximum value in the map, with a custom comparing function.

'maximumVal' == 'maximumValBy' 'compare'

minimumVal :: Ord b => Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the minimum value in the map

minimumValBy :: (b -> b -> Ordering) -> Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the minimum value in the map, with a custom comparing function.

'minimumVal' == 'minimumValBy' 'compare'

maximumValNE :: Ord b => NEMap a b -> (a, b)

Version of maximumVal for nonempty maps.

maximumValByNE :: (b -> b -> Ordering) -> NEMap a b -> (a, b)

Version of maximumValBy for nonempty maps.

minimumValNE :: Ord b => NEMap a b -> (a, b)

Version of minimumVal for nonempty maps.

minimumValByNE :: (b -> b -> Ordering) -> NEMap a b -> (a, b)

Version of minimumValBy for nonempty maps.

listTup :: [a] -> Maybe (a, a)

_ListTup :: Prism' [a] (a, a)

listTup3 :: [a] -> Maybe (a, a, a)

_ListTup3 :: Prism' [a] (a, a, a)

listTup4 :: [a] -> Maybe (a, a, a, a)

_ListTup4 :: Prism' [a] (a, a, a, a)

sortSizedBy :: Vector v a => (a -> a -> Ordering) -> Vector v n a -> Vector v n a

withAllSized :: Vector v a => NonEmpty [a] -> (forall n. KnownNat n => NonEmpty (Vector v n a) -> Maybe r) -> Maybe r

Simple type util

deleteFinite :: KnownNat n => Finite (n + 1) -> Finite (n + 1) -> Maybe (Finite n)

Delete a potential value from a Finite.

type Letter = Finite 26

charFinite :: Char -> Maybe (Bool, Finite 26)

Parse a letter into a number 0 to 25. Returns False if lowercase and True if uppercase.

_CharFinite :: Prism' Char (Bool, Finite 26)

Prism for a Char as (Bool, Finite 26), where the Finite is the letter parsed as a number from 0 to 25, and the Bool is lowercase (False) or uppercase (True).

hexDigit :: Prism' Char (Finite 16)

decimalDigit :: Prism' Char (Finite 10)

splitWord :: Word8 -> (Finite 16, Finite 16)

digitToIntSafe :: Char -> Maybe Int

caeser :: Finite 26 -> Char -> Char

Caeser shift, preserving case. If you have an Int or Integer, convert into Finite using modulo.

eitherItem :: Lens' (Either a a) a

getDown :: Down a -> a

toNatural :: Integral a => a -> Maybe Natural

Parsers

newtype TokStream a

Use a stream of tokens a as the underlying parser stream. Note that error messages for parser errors are going necessarily to be wonky.

Constructors

TokStream

Fields

getTokStream :: [a]

parseTokStream :: Foldable t => Parsec e (TokStream s) a -> t s -> Either (ParseErrorBundle (TokStream s) e) a

Parse a stream of tokens s purely, returning Either

parseTokStream_ :: (Alternative m, Foldable t) => Parsec e (TokStream s) a -> t s -> m a

Parse a stream of tokens s purely

parseTokStreamT :: (Foldable t, Monad m) => ParsecT e (TokStream s) m a -> t s -> m (Either (ParseErrorBundle (TokStream s) e) a)

Parse a stream of tokens s over an underlying monad, returning Either

parseTokStreamT_ :: (Alternative f, Foldable t, Monad m) => ParsecT e (TokStream s) m a -> t s -> m (f a)

Parse a stream of tokens s over an underlying monad

type TokParser s = Parsec Void (TokStream s)

parseWords :: Parsec Void (TokStream String) a -> String -> Maybe a

nextMatch :: MonadParsec e s m => m a -> m a

Skip every result until this token matches

parseMaybeLenient :: Parsec Void s a -> s -> Maybe a

parseOrFail :: (Stream s, ShowErrorComponent e) => Parsec e s a -> s -> a

type CharParser = Parsec Void String

pWord :: (Stream s, Token s ~ Char, Ord e) => Parsec e s String

pHWord :: (Stream s, Token s ~ Char, Ord e) => Parsec e s String

pDecimal :: (Stream s, Token s ~ Char, Ord e) => Parsec e s Int

parseLines :: Parsec Void String a -> String -> Maybe [a]

Graph

type Graph v e = Map v (Map v e)

toFGL :: (Graph gr, Ord v) => Graph v e -> (gr v e, Set v)

Points

type Point = V2 Int

2D Coordinate

cardinalNeighbs :: Point -> [Point]

cardinalNeighbsSet :: Point -> Set Point

fullNeighbs :: (Applicative f, Num a, Eq (f a), Traversable f) => f a -> [f a]

fullNeighbsSet :: (Applicative f, Num a, Ord (f a), Traversable f) => f a -> Set (f a)

mannDist :: (Foldable f, Num a, Num (f a)) => f a -> f a -> a

mulPoint :: Num a => V2 a -> V2 a -> V2 a

Treat as complex number multiplication. useful for rotations

lineTo :: Point -> Point -> [Point]

Lattice points for line between points, not including endpoints

Directions

data Dir

Constructors

North

East

South

West

parseDir :: Char -> Maybe Dir

dirPoint :: Num a => Dir -> V2 a

dirPoint' :: Num a => Dir -> V2 a

dirPoint but with inverted y axis

rotPoint :: Num a => Dir -> V2 a -> V2 a

Rotate a point by a direction

mulDir :: Dir -> Dir -> Dir

Multiply headings, taking North as straight, East as clockwise turn, West as counter-clockwise turn, and South as reverse.

Should be a commutative group; it's essentially complex number multiplication like mulPoint, with North = 1, West = i. The identity is North and the inverse is the opposite direction.

2D Maps

memoPoint :: Memo Point

boundingBox :: (Foldable1 f, Applicative g, Ord a) => f (g a) -> V2 (g a)

Find the minimum and maximum x and y from a collection of points.

Returns V2 (V2 xMin yMin) (V2 xMax yMax).

boundingBox' :: (Foldable f, Applicative g, Ord a) => f (g a) -> Maybe (V2 (g a))

A version of boundingBox that works for normal possibly-empty lists.

inBoundingBox :: (Applicative g, Foldable g, Ord a) => V2 (g a) -> g a -> Bool

parseAsciiMap :: (Char -> Maybe a) -> String -> Map Point a

asciiGrid :: IndexedTraversal Point String [a] Char a

parseAsciiSet :: (Char -> Bool) -> String -> Set Point

newtype ScanPoint

It's Point, but with a newtype wrapper so we have an Ord that sorts by y first, then x

Constructors

SP

Fields

_getSP :: Point

displayAsciiMap

Arguments

:: Char
default tile

-> Map Point Char
tile map

-> String

Recursion Schemes

anaM :: (Monad m, Corecursive t, Traversable (Base t)) => (a -> m (Base t a)) -> a -> m t

data TreeF a b

Constructors

NodeF a (ForestF a b)

type ForestF a b = [b]

Orphan instances

FoldableWithIndex k (NEMap k)

Instance details

Methods

ifoldMap :: Monoid m => (k -> a -> m) -> NEMap k a -> m

ifolded :: IndexedFold k (NEMap k a) a

ifoldr :: (k -> a -> b -> b) -> b -> NEMap k a -> b

ifoldl :: (k -> b -> a -> b) -> b -> NEMap k a -> b

ifoldr' :: (k -> a -> b -> b) -> b -> NEMap k a -> b

ifoldl' :: (k -> b -> a -> b) -> b -> NEMap k a -> b

FunctorWithIndex k (NEMap k)

Instance details

Methods

imap :: (k -> a -> b) -> NEMap k a -> NEMap k b

imapped :: IndexedSetter k (NEMap k a) (NEMap k b) a b

TraversableWithIndex k (NEMap k)

Instance details

Methods

itraverse :: Applicative f => (k -> a -> f b) -> NEMap k a -> f (NEMap k b)

itraversed :: IndexedTraversal k (NEMap k a) (NEMap k b) a b

Corecursive (Tree a)

Instance details

Methods

embed :: Base (Tree a) (Tree a) -> Tree a

ana :: (a0 -> Base (Tree a) a0) -> a0 -> Tree a

apo :: (a0 -> Base (Tree a) (Either (Tree a) a0)) -> a0 -> Tree a

postpro :: Recursive (Tree a) => (forall b. Base (Tree a) b -> Base (Tree a) b) -> (a0 -> Base (Tree a) a0) -> a0 -> Tree a

gpostpro :: (Recursive (Tree a), Monad m) => (forall b. m (Base (Tree a) b) -> Base (Tree a) (m b)) -> (forall c. Base (Tree a) c -> Base (Tree a) c) -> (a0 -> Base (Tree a) (m a0)) -> a0 -> Tree a

Recursive (Tree a)

Instance details

Methods

project :: Tree a -> Base (Tree a) (Tree a)

cata :: (Base (Tree a) a0 -> a0) -> Tree a -> a0

para :: (Base (Tree a) (Tree a, a0) -> a0) -> Tree a -> a0

gpara :: (Corecursive (Tree a), Comonad w) => (forall b. Base (Tree a) (w b) -> w (Base (Tree a) b)) -> (Base (Tree a) (EnvT (Tree a) w a0) -> a0) -> Tree a -> a0

prepro :: Corecursive (Tree a) => (forall b. Base (Tree a) b -> Base (Tree a) b) -> (Base (Tree a) a0 -> a0) -> Tree a -> a0

gprepro :: (Corecursive (Tree a), Comonad w) => (forall b. Base (Tree a) (w b) -> w (Base (Tree a) b)) -> (forall c. Base (Tree a) c -> Base (Tree a) c) -> (Base (Tree a) (w a0) -> a0) -> Tree a -> a0

(Ord k, Ord p) => Ixed (OrdPSQ k p v)

Instance details

Methods

ix :: Index (OrdPSQ k p v) -> Traversal' (OrdPSQ k p v) (IxValue (OrdPSQ k p v))

(Ixed (v a), Index (v a) ~ Int, IxValue (v a) ~ a) => Ixed (Vector v n a)

Instance details

Methods

ix :: Index (Vector v n a) -> Traversal' (Vector v n a) (IxValue (Vector v n a))