AOC.Common.Point

Points

type Point = V2 Int

2D Coordinate

type FinPoint n = V2 (Finite n)

cardinalNeighbs :: Point -> [Point]

cardinalNeighbsSet :: Point -> Set Point

fullNeighbs :: (Applicative f, Num a, Traversable f) => f a -> [f a]

fullNeighbsSet :: (Applicative f, Num a, Ord (f a), Traversable f) => f a -> Set (f a)

mannDist :: (Foldable f, Num a, Num (f a)) => f a -> f a -> a

mulPoint :: Num a => V2 a -> V2 a -> V2 a

Treat as complex number multiplication. useful for rotations

lineTo :: Point -> Point -> [Point]

Lattice points for line between points, not including endpoints

Directions

data Dir

Constructors

North

East

South

West

parseDir :: Char -> Maybe Dir

dirPoint :: Num a => Dir -> V2 a

dirPoint' :: Num a => Dir -> V2 a

dirPoint but with inverted y axis

rotPoint :: Num a => Dir -> V2 a -> V2 a

Rotate a point by a direction

rotFin :: KnownNat n => Dir -> FinPoint n -> FinPoint n

Rotate a point by a direction

mulDir :: Dir -> Dir -> Dir

Multiply headings, taking North as straight, East as clockwise turn, West as counter-clockwise turn, and South as reverse.

Should be a commutative group; it's essentially complex number multiplication like mulPoint, with North = 1, West = i. The identity is North and the inverse is the opposite direction.

allDir :: NonEmpty Dir

allDirSet :: NESet Dir

Orientations

data D8

Represents an orientation of a 2d tile.

Constructors

D8

Fields

d8Rot :: !Dir

d8Flip :: !Bool

mulD8 :: D8 -> D8 -> D8

a mulD8 b represents applying b, then a.

orientPoint :: Num a => D8 -> V2 a -> V2 a

Rotate and flip a point by a D8

orientFin :: KnownNat n => D8 -> FinPoint n -> FinPoint n

allD8 :: NonEmpty D8

allD8Set :: NESet D8

2D Maps

memoPoint :: Memo Point

boundingBox :: (Foldable1 f, Applicative g, Ord a) => f (g a) -> V2 (g a)

Find the minimum and maximum x and y from a collection of points.

Returns V2 (V2 xMin yMin) (V2 xMax yMax).

boundingBox' :: (Foldable f, Applicative g, Ord a) => f (g a) -> Maybe (V2 (g a))

A version of boundingBox that works for normal possibly-empty lists.

inBoundingBox :: (Applicative g, Foldable g, Ord a) => V2 (g a) -> g a -> Bool

minCorner :: (Foldable1 f, Applicative g, Ord a) => f (g a) -> g a

minCorner' :: (Foldable f, Applicative g, Ord a) => f (g a) -> Maybe (g a)

shiftToZero :: (Applicative f, Num a, Ord a) => NESet (f a) -> NESet (f a)

Shift corner to (0,0)

shiftToZero' :: (Applicative f, Num a, Ord a) => Set (f a) -> Set (f a)

Shift corner to (0,0)

parseAsciiMap :: (Char -> Maybe a) -> String -> Map Point a

asciiGrid :: IndexedTraversal Point String [a] Char a

parseAsciiSet :: (Char -> Bool) -> String -> Set Point

newtype ScanPoint

It's Point, but with a newtype wrapper so we have an Ord that sorts by y first, then x

Constructors

SP

Fields

_getSP :: Point

displayAsciiMap

Arguments

:: Char
default tile

-> Map Point Char
tile map

-> String

Util

centeredFinite :: forall n. KnownNat n => Iso' (Finite n) Rational