AOC.Common.Search

Documentation

aStar

Arguments

:: forall n p. (Ord n, Ord p, Num p)

=> (n -> p)
heuristic

-> (n -> Map n p)
neighborhood

-> n
start

-> (n -> Bool)
target

-> Maybe (p, [n])
the shortest path, if it exists, and its cost

A* Search

bfs

Arguments

:: forall n. Ord n

=> (n -> Set n)
neighborhood

-> n
start

-> (n -> Bool)
target

-> Maybe [n]
the shortest path, if it exists

Breadth-first search, with loop detection

binarySearch :: (Int -> Ordering) -> Int -> Int -> Maybe Int

exponentialSearch :: (Int -> Ordering) -> Int -> Maybe Int

binaryMinSearch :: (Int -> Bool) -> Int -> Int -> Maybe Int

Find the lowest value where the predicate is satisfied within the given bounds.

exponentialMinSearch :: (Int -> Bool) -> Int -> Maybe Int

Find the lowest value where the predicate is satisfied above a given bound.

binaryFindMin :: (Int -> Maybe a) -> Int -> Int -> Maybe a

Find the lowest value where the predicate is Just within the given bounds.

exponentialFindMin :: (Int -> Maybe a) -> Int -> Maybe a

Find the lowest value where the predicate is Just above a given bound.