AOC.Discover

Description

Template Haskell for discovering all named challenges in a given directory.

Documentation

mkChallengeMap :: [(Day, (Part, SomeSolution))] -> ChallengeMap

Meant to be called like:

mkChallengeMap $$(solutionList "src/AOC/Challenge")

solutionList :: FilePath -> Q (TExp [(Day, (Part, SomeSolution))])

Template Haskell splice to produce a list of all named solutions in a directory. Expects solutions as function names following the format dayDDp, where DD is a two-digit zero-added day, and p is a lower-case letter corresponding to the part of the challenge.

See mkChallengeMap for a description of usage.

type ChallengeMap = Map Day (Map Part SomeSolution)

A map of days to parts to solutions.

data ChallengeSpec

A specification for a specific challenge. Should consist of a day and a lowercase character.

Constructors

CS

Fields

_csDay :: Day

_csPart :: Part

solSpec :: Name -> ChallengeSpec

Get a ChallengeSpec from a given reified solution (name).

solSpec 'day02a == CS { _csDay = 1, _csPart = a }

solSpecStr :: String -> Either (ParseErrorBundle String Void) ChallengeSpec

solSpecStr_ :: String -> ChallengeSpec

charPart :: Char -> Maybe Part

Parse a Char into a Part

challengeName :: Parser ChallengeSpec

solverNFData :: Name -> Q Bool

Check if a solver identifier is of type A :~> B, where B is an instance of NFData.

deepInstance

Arguments

:: Name
class

-> Type
type

-> Q Bool

Check if a type is an instance of a class, unifying when possible