AOC.Run.Interactive

Description

Versions of loaders and runners meant to be used in GHCI.

Fetch and Run

Return Answers

execSolution :: ChallengeSpec -> IO String

Run the solution indicated by the challenge spec on the official puzzle input. Get answer as result.

execSolutionWith

Arguments

:: ChallengeSpec

-> String
custom puzzle input

-> IO String

Run the solution indicated by the challenge spec on a custom input. Get answer as result.

testSolution :: ChallengeSpec -> IO (Maybe Bool)

Run test suite for a given challenge spec.

Returns Just if any tests were run, with a Bool specifying whether or not all tests passed.

viewPrompt :: ChallengeSpec -> IO Text

View the prompt for a given challenge spec.

waitForPrompt :: ChallengeSpec -> IO Text

Countdown to get the prompt for a given challenge spec, if not yet available.

submitSolution :: ChallengeSpec -> IO (Text, SubmitRes)

Submit solution for a given challenge spec, lock if correct, and retry if a wait is received.

No Answers

execSolution_ :: ChallengeSpec -> IO ()

Result-suppressing version of execSolution.

execSolutionWith_

Arguments

:: ChallengeSpec

-> String
custom puzzle input

-> IO ()

Result-suppressing version of execSolutionWith.

testSolution_ :: ChallengeSpec -> IO ()

Result-suppressing version of testSolution.

viewPrompt_ :: ChallengeSpec -> IO ()

Result-suppressing version of viewPrompt.

waitForPrompt_ :: ChallengeSpec -> IO ()

Result-suppressing version of waitForPrompt.

submitSolution_ :: ChallengeSpec -> IO ()

Result-suppressing version of submitSolution.

Load Inputs

loadInput :: ChallengeSpec -> IO String

Load input for a given challenge

loadParseInput :: ChallengeSpec -> (a :~> b) -> IO a

Run the parser of a solution, given its ChallengeSpec.

loadParseInput (solSpec 'day01a) day01a

loadTests :: ChallengeSpec -> IO [(String, TestMeta)]

Load test cases for a given challenge

loadParseTests :: ChallengeSpec -> (a :~> b) -> IO [(Maybe a, TestMeta)]

Run the parser of a solution on test data, given its ChallengeSpec.

loadParseTests (solSpec 'day01a) day01a

Util

mkSpec :: Integer -> Part -> ChallengeSpec

Unsafely create a ChallengeSpec from a day number and part.

Is undefined if given a day number out of range (1-25).