AOC.Solver

Description

Types to drive the challenge runner and help speed up/clean up solutions.

Documentation

data a :~> b

Abstracting over the type of a challenge solver to help with cleaner solutions.

A a :~> b encapsulates something that solves a challenge with input type a into a response of type b.

Consists of a parser, a shower, and a solver. The solver solves a general a -> Maybe b function, and the parser and shower are used to handle the boilerplate of parsing and printing the solution.

Constructors

MkSol

Fields

sParse :: String -> Maybe a

parse input into an a

sSolve :: (?dyno :: DynoMap) => a -> Maybe b

solve an a input to a b solution

sShow :: b -> String

print out the b solution in a pretty way

withSolver :: (String -> Maybe String) -> String :~> String

Construct a :~> from a String -> Maybe String solver, which might fail. Does no parsing or special printing treatment.

withSolver' :: (String -> String) -> String :~> String

Construct a :~> from just a normal String -> String solver. Does no parsing or special printing treatment.

data SomeSolution where

Wrap an a :~> b and hide the type variables so we can put different solutions in a container.

Constructors

MkSomeSolWH :: (a :~> b) -> SomeSolution

MkSomeSolNF :: (NFData a, NFData b) => (a :~> b) -> SomeSolution

Bundled Patterns

pattern MkSomeSol :: forall a b. (a :~> b) -> SomeSolution
Handy pattern to work with both MkSomeSolWH and MkSomeSolNF. As a constructor, just uses MkSomeSolWH, so might not be desirable.

data SolutionError

Errors that might happen when running a :~> on some input.

Constructors

SEParse

SESolve

runSolution :: (a :~> b) -> String -> Either SolutionError String

Run a :~> on some input.

runSomeSolution :: SomeSolution -> String -> Either SolutionError String

Run a SomeSolution on some input.

ssIsNF :: SomeSolution -> Bool

Check if a SomeSolution is equipped with an NFData instance on the types

DynoMap

runSolutionWith

Arguments

:: Map String Dynamic
map of dynamic values for testing with lookupDyno.

-> (a :~> b)

-> String

-> Either SolutionError String

Run a :~> on some input, with a map of dynamic values for testing

runSomeSolutionWith

Arguments

:: Map String Dynamic
map of dynamic values for testing with lookupDyno.

-> SomeSolution

-> String

-> Either SolutionError String

Run a SomeSolution on some input, with a map of dynamic values for testing

dyno :: forall a. (Typeable a, ?dyno :: DynoMap) => String -> Maybe a

From a ?dyno Implicit Params, look up a value at a given key. Meant to be used with TypeApplications:

'dyno' @"hello"

This can be used within the body of sSolve, since it will always be called with the implicit parameter.

When called on actual puzzle input, result will always be Nothing. But, for some test inputs, there might be supplied values.

This is useful for when some problems have parameters that are different with test inputs than for actual inputs.

dyno_

Arguments

:: forall a. (Typeable a, ?dyno :: DynoMap)

=> String

-> a
default

-> a

A version of dyno taking a default value in case the key is not in the map. When called on actual puzzle input, this is always id. However, for some test inputs, there might be supplied values.

Meant to be used with TypeApplications:

'dyno_' @"hello" 7

This is useful for when some problems have parameters that are different with test inputs than for actual inputs.