AOC.Util

Description

Assorted utility functions and orphans used for solutions.

Documentation

strip :: String -> String

Strip trailing and leading whitespace.

stripNewline :: String -> String

Strip trailing newline

eitherToMaybe :: Alternative m => Either e a -> m a

Convert an Either into a Maybe, or any Alternative instance, forgetting the error value.

firstJust :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Like find, but instead of taking an a -> Bool, takes an a -> Maybe b and returns the first success.

maybeToEither :: MonadError e m => e -> Maybe a -> m a

Convert a Maybe into an Either, or any MonadError instance, by providing an error value in case Nothing was given.

maybeAlt :: Alternative m => Maybe a -> m a

Generalize a Maybe to any Alternative

traceShowIdMsg :: Show a => String -> a -> a

Like traceShowId but with an extra message

traceShowMsg :: Show a => String -> a -> b -> b

Like traceShow but with an extra message