AOC.Util.DynoMap

Documentation

newtype DynoMap

Constructors

Dyno

Fields

runDyno :: Map String Dynamic

lookupDyno :: forall a. Typeable a => String -> DynoMap -> Maybe a

Lookup the value at a given key in a Dyno.

lookupDyno "hello"

lookupDynoWith :: forall a. Typeable a => String -> a -> DynoMap -> a

Like lookupDyno, but with a default value to be returned if the key is not found or has the wrong type.